Port to Python 3.10 + native mujoco 2.2.1
Compatibility fixes to run METRA on Python 3.10 with native mujoco
bindings (replaces mujoco-py which only supports Python ≤ 3.8).
Core algorithm code (iod/) is unchanged.

envs/mujoco/mujoco_utils.py:
- Add _SimProxy class and sim property to MujocoTrait
  Bridges mujoco-py style self.sim.data/model API to native mujoco
  gym attributes self.data / self.model (gym >= 0.22)

garagei/envs/consistent_normalized_env.py:
- Cast observations to float32 after normalize and flatten
  gym 0.24 sets observation_space dtype=float64; gym.spaces.utils.flatten
  preserves that dtype. PyTorch 2.0 F.linear requires identical dtypes
  for input and weights (float32), so explicit casts are needed here.

garaged/src/garage/tf/__init__.py:
garaged/src/garage/tf/misc/tensor_utils.py:
garaged/src/garage/tf/samplers/__init__.py:
- Wrap tensorflow imports in try/except ImportError
  TensorFlow is not installed (and not needed for METRA/SAC). These
  modules are transitively imported at startup via garage.np.algos.

garaged/src/garage/np/algos/cem.py:
garaged/src/garage/np/algos/cma_es.py:
- Make BatchSampler import lazy (inside _get_batch_sampler())
  Avoids triggering the tensorflow import chain at module load time.

requirements.txt:
- Comment out install==1.3.5 (unavailable on PyPI, not needed)

+++ garagei/envs/consistent_normalized_env.py
@@ -40,7 +40,7 @@ class ConsistentNormalizedEnv(AkroWrapperTrait, gym.Wrapper):
 
     def _apply_normalize_obs(self, obs):
         normalized_obs = (obs - self._obs_mean) / (np.sqrt(self._obs_var) + 1e-8)
-        return normalized_obs
+        return normalized_obs.astype(np.float32)
 
     def reset(self, **kwargs):
         obs = self.env.reset(**kwargs)
@@ -51,6 +51,7 @@ class ConsistentNormalizedEnv(AkroWrapperTrait, gym.Wrapper):
 
         if self._flatten_obs:
             obs = gym.spaces.utils.flatten(self.env.observation_space, obs)
+            obs = obs.astype(np.float32)
 
         return obs
 
@@ -78,6 +79,7 @@ class ConsistentNormalizedEnv(AkroWrapperTrait, gym.Wrapper):
 
         if self._flatten_obs:
             next_obs = gym.spaces.utils.flatten(self.env.observation_space, next_obs)
+            next_obs = next_obs.astype(np.float32)
 
         return next_obs, reward, done, info
